+++ src/mopidy/backend.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 import logging
+from collections.abc import Iterable
 from typing import TYPE_CHECKING, Any, ClassVar
 
 import pykka
@@ -137,10 +138,21 @@ class LibraryProvider:
         """
         return {}
 
-    def lookup(self, uri: Uri) -> list[Track]:
+    def lookup_many(self, uris: Iterable[Uri]) -> dict[Uri, list[Track]]:
         """See :meth:`mopidy.core.LibraryController.lookup`.
 
         *MUST be implemented by subclass.*
+        """
+        return {uri: self.lookup(uri) for uri in uris}
+
+    def lookup(self, uri: Uri) -> list[Track]:
+        """See :meth:`mopidy.core.LibraryController.lookup`.
+
+        *MUST be implemented by subclass if :meth:`lookup_many` is not implemented.*
+
+        .. deprecated:: 4.0
+            Implement :meth:`lookup_many` instead. If :meth:`lookup_many` is
+            implemented, Mopidy will never call this method on a backend.
         """
         raise NotImplementedError
 
@@ -467,7 +479,7 @@ class LibraryProviderProxy:
     browse = proxy_method(LibraryProvider.browse)
     get_distinct = proxy_method(LibraryProvider.get_distinct)
     get_images = proxy_method(LibraryProvider.get_images)
-    lookup = proxy_method(LibraryProvider.lookup)
+    lookup_many = proxy_method(LibraryProvider.lookup_many)
     refresh = proxy_method(LibraryProvider.refresh)
     search = proxy_method(LibraryProvider.search)