+++ docs/changelog.rst
@@ -82,7 +82,13 @@ Backend API
 
 Changes to the Backend API may affect Mopidy backend extensions.
 
-- No changes so far.
+- Added :meth:`mopidy.backend.LibraryProvider.lookup_many` to take a list of
+  URIs and return a mapping of URIs to tracks. If this method is not implemented
+  then repeated calls to :meth:`mopidy.backend.LibraryProvider.lookup` will be
+  used as a fallback.
+
+- Deprecated :meth:`mopidy.backend.LibraryProvider.lookup`. Extensions should
+  implement :meth:`mopidy.backend.LibraryProvider.lookup_many` instead.
 
 Models
 ------
@@ -116,6 +122,9 @@ Extension support
 
 Internals
 ---------
+
+- Dropped split between the ``main`` and ``develop`` branches. We now use
+  ``main`` for all development, and have removed the ``develop`` branch.
 
 - Added type hints to most of the source code.